angle between two vectors

+++ tinydot/lib.py
@@ -36,7 +36,7 @@ class LIB(metaclass=Singleton):
     self.v_dot   = self.c_wrapper('vector_dot',         c_double,  [c_void_p, c_void_p])
     self.v_sub   = self.c_wrapper('vector_sub',         c_void_p,  [c_void_p, c_void_p])
     self.v_norm  = self.c_wrapper('vector_norm',        c_double,  [c_void_p])
-    self.v_angle = self.c_wrapper('vector_angle',       c_double,  [c_void_p, c_void_p, c_double])
+    self.v_angle = self.c_wrapper('vector_angle',       c_double,  [c_void_p, c_void_p, c_uint])
     self.v_rotate= self.c_wrapper('vector_rotate',      c_void_p,  [c_void_p, c_double])
 
   def c_wrapper(self, funcname, restype, argtypes):
